Deployment-Scripts/requirements.txt
requests
PyGithub
#github.py
#boto3